requirements.txt
django==1.7.1
psycopg2==2.5.4
python-dateutil
